Some refactoring and code moving.
