+++ graphviz/dr2xml_total_dot.txt
@@ -9,7 +9,7 @@ subgraph cluster0{label="Module dr2xml_total.py";
                   n003 [label = "analyze_priority"];
                   n004 [label = "wr"];
                   n005 [label = "freq2datefmt"];
-                  n006 [label = "write_xios_file_def"];
+                  n006 [label = "write_xios_file_def_for_svar"];
                   n007 [label = "wrv"];
                   n008 [label = "create_xios_aux_elmts_defs"];
                   n009 [label = "gather_AllSimpleVars"];
@@ -92,7 +92,7 @@ subgraph cluster3{label="Class simple_Dim ";
                   n008 -> n016 /*analyze_cell_time_method*/;
                   n009 -> n002 /*select_CMORvars_for_lab*/;
                   n009 -> n033 /*process_homeVars*/;
-                  n010 -> n006 /*write_xios_file_def*/;
+                  n010 -> n006 /*write_xios_file_def_for_svar*/;
                   n010 -> n009 /*gather_AllSimpleVars*/;
                   n010 -> n011 /*print_SomeStats*/;
                   n010 -> n014 /*create_xios_axis_and_grids_for_plevs_unions*/;
